Move ctypes into their own module
